chore : review app 	- add v1 	- imporove v2 : outlier handler, using miceforest instead Iterative Inputer
