+++ requirements.txt
@@ -2,7 +2,7 @@ streamlit
 pandas
 numpy
 matplotlib
+seaborn
 scikit-learn
 miceforest
-seaborn
-scipy
+openpyxl
